backend/requirements.txt
alembic==1.17.1
black==25.11.0 # formating for alembic
ddgs==9.9.0
duckduckgo-search==8.1.1
fastapi==0.121.1
httpx==0.28.1
logfire==4.14.2
openai==2.7.2
pydantic==2.12.4
pydantic-ai==1.9.1
pydantic-settings==2.12.0
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
psycopg==3.2.12
psycopg-binary==3.2.12
sqlalchemy==2.0.44
typing-extensions==4.15.0
uvicorn==0.38.0